generalize api method
- (Program.cs) allow using api from anywhere(public api).
- (ActController) improve the method to support various input.

+++ homework3/ToDoAPI/Controllers/ActivitiesController.cs
@@ -23,12 +23,21 @@ public class ActivitiesController : ControllerBase
     public IActionResult Get()
     {
         ToDoDbContext db = new ToDoDbContext();
-        string userName = getUserName();
+        string userId;
+        try
+        {
+            userId = getUserId();
+        }
+        catch (Exception)
+        {
+            return StatusCode(403, new { detail = "can't get user id from the token" }); //forbid
+        }
 
-        IQueryable<Activity> activities = from a in db.Activity select a;
-        activities = activities.Where(act => act.UserId.Equals(userName));
+        IQueryable<Activity> activities = from act in db.Activity
+                                          where act.UserId.Equals(userId)
+                                          select act;
+
         if (!activities.Any()) return NoContent();
-
 
         return Ok(activities);
     }
@@ -38,28 +47,45 @@ public class ActivitiesController : ControllerBase
     [Authorize(Roles = "user")]
     public IActionResult Get(uint id)
     {
-        ToDoDbContext db = new ToDoDbContext();
-        string userName = getUserName();
-        IQueryable<Activity> activities = from a in db.Activity select a;
-        activities = activities.Where(act => act.UserId.Equals(userName));
-        activities = activities.Where(act => act.Id.Equals(id));
+        ToDoDbContext db = new();
+        string userId;
+        try
+        {
+            userId = getUserId();
+        }
+        catch (Exception)
+        {
+            return StatusCode(403, new { detail = "can't get user id from the token" }); //forbid
+        }
+        IQueryable<Activity> activities = from act in db.Activity
+                                          where act.Id == id && act.UserId.Equals(userId)
+                                          select act;
 
-        if (activities == null) return NotFound(new { detail = "can't find activity id: " + id });
+        Activity? activity = activities.FirstOrDefault();
+        if (activity == null) return NotFound(new { detail = "can't find activity id: " + id });
 
-        return Ok(activities.FirstOrDefault());
+        return Ok(activity);
     }
 
     [HttpPost]
     [Authorize(Roles = "user")]
     public IActionResult Post([FromBody] DTOs.Activity data)
     {
-        ToDoDbContext db = new ToDoDbContext();
-        string userName = getUserName();
+        ToDoDbContext db = new();
+        string userId;
+        try
+        {
+            userId = getUserId();
+        }
+        catch (Exception)
+        {
+            return StatusCode(403, new { detail = "can't get user id from the token" }); //forbid
+        }
 
         Activity activity = new Models.Activity();
         activity.Name = data.name;
         activity.Appoint = data.appoint;
-        activity.UserId = userName;
+        activity.UserId = userId;
 
         db.Activity.Add(activity);
         db.SaveChanges();
@@ -72,15 +98,25 @@ public class ActivitiesController : ControllerBase
     [Authorize(Roles = "user")]
     public IActionResult Put(uint id, [FromBody] DTOs.Activity data)
     {
-        ToDoDbContext db = new ToDoDbContext();
-        string userName = getUserName();
-        IQueryable<Activity> activities = from a in db.Activity select a;
-        activities = activities.Where(act => act.UserId.Equals(userName));
-        activities = activities.Where(act => act.Id.Equals(id));
-        Activity activity = activities.FirstOrDefault();
+        ToDoDbContext db = new();
+        string userId;
+        try
+        {
+            userId = getUserId();
+        }
+        catch (Exception)
+        {
+            return StatusCode(403, new { detail = "can't get user id from the token" }); //forbid
+        }
+
+        IQueryable<Activity> activities = from act in db.Activity
+                                          where act.Id == id && act.UserId.Equals(userId)
+                                          select act;
+        Activity? activity = activities.FirstOrDefault();
+        if (activity == null) return NotFound(new { detail = "can't find the activity" });
+
         activity.Name = data.name;
         activity.Appoint = data.appoint;
-
         db.SaveChanges();
 
         return Ok(new { detail = "Activity updated successfully." });
@@ -91,13 +127,22 @@ public class ActivitiesController : ControllerBase
     [Authorize(Roles = "user")]
     public IActionResult Delete(uint id)
     {
-        ToDoDbContext db = new ToDoDbContext();
-        string userName = getUserName();
-        IQueryable<Activity> activities = from a in db.Activity select a;
-        activities = activities.Where(act => act.UserId.Equals(userName));
-        activities = activities.Where(act => act.Id.Equals(id));
+        ToDoDbContext db = new();
+        string userId;
+        try
+        {
+            userId = getUserId();
+        }
+        catch (Exception)
+        {
+            return StatusCode(403, new { detail = "can't get user id from the token" }); //forbid
+        }
 
-        Activity activity = activities.FirstOrDefault();
+        IQueryable<Activity> activities = from act in db.Activity
+                                          where act.Id == id && act.UserId.Equals(userId)
+                                          select act;
+        Activity? activity = activities.FirstOrDefault();
+        if (activity == null) return NotFound(new { detail = "can't find the activity" });
 
         db.Activity.Remove(activity);
         db.SaveChanges();
@@ -105,11 +150,12 @@ public class ActivitiesController : ControllerBase
         return Ok(new { detail = "Activity deleted successfully." });
     }
 
-    private string getUserName()
+    private string getUserId()
     {
         // Get the Authorization header from the HTTP request
-        ClaimsIdentity identity = HttpContext.User.Identity as ClaimsIdentity;
-        var a = identity.FindFirst(ClaimTypes.Name).Value;
-        return a;
+        if (HttpContext.User.Identity is not ClaimsIdentity identity) throw new Exception("Claims identity not found");
+
+        Claim? claim = identity.FindFirst(ClaimTypes.Name) ?? throw new Exception("The identity has no claim");
+        return claim.Value;
     }
 }
